requirements.txt
qrcode[pil]
PyQt6
Pillow
qrcode[svg]
